+++ src/components/form/checkbox.vue
@@ -14,6 +14,7 @@
 
 <script>
 import '../../material-design-lite/checkbox/checkbox';
+import '../../material-design-lite/ripple/ripple';
 
 const CALLBACK_CHANGE = 'change';
 
@@ -81,6 +82,9 @@ export default {
   },
   mounted() {
     this.$ui.upgradeElement(this.$el, 'MaterialCheckbox');
+    if (this.effect) {
+      this.$ui.upgradeElement(this.$el, 'MaterialRipple');
+    }
   }
 };
 </script>