+++ src/components/form/radio.vue
@@ -14,6 +14,7 @@
 
 <script>
 import '../../material-design-lite/radio/radio';
+import '../../material-design-lite/ripple/ripple';
 
 const CALLBACK_CHANGE = 'change';
 
@@ -70,6 +71,9 @@ export default {
   },
   mounted() {
     this.$ui.upgradeElement(this.$el, 'MaterialRadio');
+    if (this.effect) {
+      this.$ui.upgradeElement(this.$el, 'MaterialRipple');
+    }
   }
 };
 </script>